fix: break sdk/patches circular import by moving is_tracing_armed to a leaf module
