+++ src/traceml_ai/instrumentation/patches/dataloader_patch.py
@@ -1,6 +1,6 @@
 from torch.utils.data import DataLoader
 
-from traceml_ai.sdk.initial import is_tracing_armed
+from traceml_ai.runtime.arming import is_tracing_armed
 from traceml_ai.utils.timing import timed_region
 
 _ORIG_DATALOADER_ITER = DataLoader.__iter__
