+++ src/traceml_ai/instrumentation/patches/h2d_auto_timer_patch.py
@@ -46,7 +46,7 @@ from typing import Any
 import torch
 
 from traceml_ai.instrumentation.h2d import should_time_h2d
-from traceml_ai.sdk.initial import is_tracing_armed
+from traceml_ai.runtime.arming import is_tracing_armed
 from traceml_ai.utils.timing import timed_region
 
 _H2D_TLS = threading.local()
